flask/requirements.txt
gunicorn>=21.2.0
Flask>=3.0.0
Flask-Session>=0.6.0
flask_login>=0.6.3
flask_sqlalchemy>=3.1.1
SQLAlchemy>=2.0.25
Werkzeug>=3.0.1
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
flask_wtf>=1.2.1
WTForms>=3.1.2
